Added SubscriberFileProcessor class and Interface. Refactored tests accordingly.
