+++ src/Platform.Eda.Cli/Commands/ConfigureEda/JsonProcessor/ISubscriberFileParser.cs
@@ -1,0 +1,10 @@
+﻿using System.Text;
+using Platform.Eda.Cli.Commands.ConfigureEda.Models;
+
+namespace Platform.Eda.Cli.Commands.ConfigureEda.JsonProcessor
+{
+    public interface ISubscriberFileParser
+    {
+        public PutSubscriberFile ParseFile(string fileName);
+    }
+}
