+++ src/Platform.Eda.Cli/Commands/ConfigureEda/JsonProcessor/SubscribersDirectoryProcessor.cs
@@ -4,8 +4,6 @@ using System.IO;
 using System.IO.Abstractions;
 using System.Linq;
 using CaptainHook.Domain.Results;
-using Newtonsoft.Json;
-using Newtonsoft.Json.Linq;
 using Platform.Eda.Cli.Commands.ConfigureEda.Models;
 using Platform.Eda.Cli.Common;
 
@@ -14,10 +12,12 @@ namespace Platform.Eda.Cli.Commands.ConfigureEda.JsonProcessor
     public class SubscribersDirectoryProcessor : ISubscribersDirectoryProcessor
     {
         private readonly IFileSystem _fileSystem;
+        private readonly ISubscriberFileParser _subscriberFileParser;
 
-        public SubscribersDirectoryProcessor(IFileSystem fileSystem)
+        public SubscribersDirectoryProcessor(IFileSystem fileSystem, ISubscriberFileParser subscriberFileParser)
         {
             _fileSystem = fileSystem;
+            _subscriberFileParser = subscriberFileParser;
         }
 
         public OperationResult<IEnumerable<PutSubscriberFile>> ProcessDirectory(string inputFolderPath)
@@ -33,7 +33,7 @@ namespace Platform.Eda.Cli.Commands.ConfigureEda.JsonProcessor
             try
             {
                 var subscriberFiles = _fileSystem.Directory.GetFiles(sourceFolderPath, "*.json", SearchOption.AllDirectories);
-                subscribers.AddRange(subscriberFiles.Select(ProcessFile));
+                subscribers.AddRange(subscriberFiles.Select(_subscriberFileParser.ParseFile));
 
                 if (!subscribers.Any())
                 {
@@ -46,66 +46,6 @@ namespace Platform.Eda.Cli.Commands.ConfigureEda.JsonProcessor
             }
 
             return subscribers;
-        }
-
-        private PutSubscriberFile ProcessFile(string fileName)
-        {
-            try
-            {
-                var content = _fileSystem.File.ReadAllText(fileName);
-                var contentJObject = JObject.Parse(content);
-                var varsDictionary = GetFileVars(contentJObject);
-
-                var request = JsonConvert.DeserializeObject<PutSubscriberRequest>(contentJObject.ToString());
-
-                return new PutSubscriberFile
-                {
-                    File = new FileInfo(fileName),
-                    Request = request
-                };
-            }
-            catch (Exception ex)
-            {
-                return new PutSubscriberFile
-                {
-                    File = new FileInfo(fileName),
-                    Error = ex.Message
-                };
-            }
-        }
-
-        public static Dictionary<string, Dictionary<string, string>> GetFileVars(JObject fileContent)
-        {
-            if (fileContent.ContainsKey("vars"))
-            {
-                var varsDict = JsonConvert.DeserializeObject<Dictionary<string, Dictionary<string, JToken>>>(fileContent["vars"]!.ToString());
-
-                var outputVarsDict = new Dictionary<string, Dictionary<string, string>>();
-
-                foreach (var (propertyKey, innerDict) in varsDict)
-                {
-                    outputVarsDict[propertyKey] = new Dictionary<string, string>();
-                    foreach (var (envKey, val) in innerDict)
-                    {
-                        var stringVal = val.Type == JTokenType.String ? val.ToString() : val.ToString(Formatting.None);
-                        if (envKey.Contains(","))
-                        {
-                            foreach (var singleEnv in envKey.Split(','))
-                            {
-                                outputVarsDict[propertyKey][singleEnv] = stringVal;
-                            }
-                        }
-                        else // convert to string
-                        {
-                            outputVarsDict[propertyKey][envKey] = stringVal;
-                        }
-                    }
-                }
-
-                fileContent.Remove("vars");
-                return outputVarsDict;
-            }
-            return new Dictionary<string, Dictionary<string, string>>(); // no vars
         }
     }
 }
