+++ src/Tests/Platform.Eda.Cli.Tests/Commands/ConfigureEda/SubscribersDirectoryProcessorTests.cs
@@ -1,11 +1,13 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.IO;
 using System.IO.Abstractions;
 using System.IO.Abstractions.TestingHelpers;
 using CaptainHook.Api.Client.Models;
 using CaptainHook.Domain.Results;
 using Eshopworld.Tests.Core;
 using FluentAssertions;
+using Moq;
 using Newtonsoft.Json.Linq;
 using Platform.Eda.Cli.Commands.ConfigureEda.JsonProcessor;
 using Platform.Eda.Cli.Commands.ConfigureEda.Models;
@@ -18,10 +20,13 @@ namespace Platform.Eda.Cli.Tests.Commands.ConfigureEda
     {
         private const string MockCurrentDirectory = @"Z:\Sample\";
         private readonly SubscribersDirectoryProcessor _subscribersDirectoryProcessor;
+        private readonly Mock<ISubscriberFileParser> _mockSubscriberFileParser;
 
         public SubscribersDirectoryProcessorTests()
         {
-            _subscribersDirectoryProcessor = new SubscribersDirectoryProcessor(new MockFileSystem(GetOneSampleInputFile(), MockCurrentDirectory));
+            var mockFs = new MockFileSystem(GetOneSampleInputFile(), MockCurrentDirectory);
+            _mockSubscriberFileParser = new Mock<ISubscriberFileParser>();
+            _subscribersDirectoryProcessor = new SubscribersDirectoryProcessor(mockFs, _mockSubscriberFileParser.Object);
         }
 
         [Theory, IsUnit]
@@ -42,25 +47,30 @@ namespace Platform.Eda.Cli.Tests.Commands.ConfigureEda
             result.Error.Should().BeOfType<CliExecutionError>();
         }
 
-        [Theory, IsUnit]
-        [InlineData(MockCurrentDirectory)]
-        public void ProcessDirectory_WithValidFile_ReturnsValidObject(string testFilesDirectoryPath)
+        [Fact, IsUnit]
+        public void ProcessDirectory_WithValidFile_ReturnsValidObject()
         {
-            // Act
-            var result = _subscribersDirectoryProcessor.ProcessDirectory(testFilesDirectoryPath);
+            // Arrange
+            var parsedSubFile = new PutSubscriberFile
+            {
+                Request = new PutSubscriberRequest
+                {
+                    SubscriberName = "test-sub",
+                    EventName = "test-event"
+                }
+            };
 
+            _mockSubscriberFileParser.Setup(parser => parser.ParseFile(It.IsAny<string>()))
+                .Returns(parsedSubFile);
+
+            // Act
+            var result = _subscribersDirectoryProcessor.ProcessDirectory(MockCurrentDirectory);
+            
             // Assert
             var expected = new OperationResult<IEnumerable<PutSubscriberFile>>(
                 new[]
                 {
-                    new PutSubscriberFile
-                    {
-                        Request = new PutSubscriberRequest
-                        {
-                            SubscriberName = "test-sub",
-                            EventName = "test-event",
-                        }
-                    }
+                    parsedSubFile
                 });
 
             result.Should().BeEquivalentTo(expected, opt => opt
@@ -69,182 +79,40 @@ namespace Platform.Eda.Cli.Tests.Commands.ConfigureEda
 
         }
 
-        [Theory, IsUnit]
-        [InlineData(MockCurrentDirectory)]
-        public void ProcessDirectory_WithMultipleValidFiles_ReturnsValidObject(string testFilesDirectoryPath)
+        [Fact, IsUnit]
+        public void ProcessDirectory_WithMultipleValidFiles_FileParserIsCalledForAll()
         {
-            var subscribersDirectoryProcessor = new SubscribersDirectoryProcessor(new MockFileSystem(GetMultipleMockInputFiles(), MockCurrentDirectory));
+            var mockFileSystem = new MockFileSystem(GetMultipleMockInputFiles(), MockCurrentDirectory);
+            var mockSubscriberFileParser = new Mock<ISubscriberFileParser>();
+            var subscribersDirectoryProcessor = new SubscribersDirectoryProcessor(mockFileSystem, mockSubscriberFileParser.Object);
+
+            mockSubscriberFileParser.Setup(parser => parser.ParseFile(It.IsAny<string>()))
+                .Returns(new PutSubscriberFile());
+
             // Act
-            var result = subscribersDirectoryProcessor.ProcessDirectory(testFilesDirectoryPath);
+            var result = subscribersDirectoryProcessor.ProcessDirectory(MockCurrentDirectory);
 
             // Assert
-            var expected = new OperationResult<IEnumerable<PutSubscriberFile>>(
-                new[]
-                {
-                    new PutSubscriberFile
-                    {
-                        Request = new PutSubscriberRequest
-                        {
-                            SubscriberName = "test-sub",
-                            EventName = "test-event",
-                        }
-                    },
-                    new PutSubscriberFile
-                    {
-                        Request = new PutSubscriberRequest
-                        {
-                            SubscriberName = "test-sub2",
-                            EventName = "test-event2",
-                        }
-                    }
-                });
-
-            result.Should().BeEquivalentTo(expected, opt => opt
-                .Excluding(info => info.SelectedMemberInfo.MemberType == typeof(CaptainHookContractSubscriberDto))
-                .Excluding(info => info.SelectedMemberInfo.MemberType == typeof(FileInfoBase)));
-
+            mockSubscriberFileParser.Verify(parser => parser.ParseFile(Path.Combine(MockCurrentDirectory, "sample1.json")), Times.Once);
+            mockSubscriberFileParser.Verify(parser => parser.ParseFile(Path.Combine(MockCurrentDirectory, "subdir\\sample2.json")), Times.Once);
+            result.IsError.Should().BeFalse();
         }
 
-        [Theory, IsUnit]
-        [InlineData(MockCurrentDirectory)]
-        public void ProcessDirectory_EmptyDirectory_ReturnsValidObject(string testFilesDirectoryPath)
+        [Fact, IsUnit]
+        public void ProcessDirectory_EmptyDirectory_ReturnsValidObject()
         {
             // Arrange
             var subscribersDirectoryProcessor = new SubscribersDirectoryProcessor(
-                new MockFileSystem(new Dictionary<string, MockFileData>(), MockCurrentDirectory));
+                new MockFileSystem(new Dictionary<string, MockFileData>(), MockCurrentDirectory), Mock.Of<ISubscriberFileParser>());
 
             // Act
-            var result = subscribersDirectoryProcessor.ProcessDirectory(testFilesDirectoryPath);
+            var result = subscribersDirectoryProcessor.ProcessDirectory(MockCurrentDirectory);
 
             // Assert
             var expected = new OperationResult<IEnumerable<PutSubscriberFile>>(new CliExecutionError("No files have been found in 'Z:\\Sample\\'"));
 
             result.Should().BeEquivalentTo(expected);
 
-        }
-
-        [Fact, IsUnit]
-        public void ProcessDirectory_WithInvalidJsonFile_ReturnsCollectionWithErrorInformation()
-        {
-            var subscribersDirectoryProcessor = new SubscribersDirectoryProcessor(new MockFileSystem(GetInvalidJsonMockInputFile(), MockCurrentDirectory));
-            // Act
-            var result = subscribersDirectoryProcessor.ProcessDirectory(MockCurrentDirectory);
-
-            // Assert
-            var expected = new OperationResult<IEnumerable<PutSubscriberFile>>(
-                new[]
-                {
-                    new PutSubscriberFile
-                    {
-                        Request = null,
-                        Error = "Unexpected character encountered while parsing value: <. Path '', line 0, position 0."
-                    }
-                });
-
-            result.Should().BeEquivalentTo(expected, opt => opt
-                .Excluding(info => info.SelectedMemberInfo.MemberType == typeof(CaptainHookContractSubscriberDto))
-                .Excluding(info => info.SelectedMemberInfo.MemberType == typeof(FileInfoBase)));
-        }
-
-        [Fact, IsUnit]
-        public void ProcessDirectory_JsonFileWithCallbacksDlqHooks_ReturnsValidObject()
-        {
-            var subscribersDirectoryProcessor = new SubscribersDirectoryProcessor(new MockFileSystem(GetSampleInputFileWithDlq(), MockCurrentDirectory));
-            // Act
-            var result = subscribersDirectoryProcessor.ProcessDirectory(MockCurrentDirectory);
-
-            // Assert
-            var expected = new OperationResult<IEnumerable<PutSubscriberFile>>(
-                new[]
-                {
-                    new PutSubscriberFile
-                    {
-                        Request = new PutSubscriberRequest
-                        {
-                            SubscriberName = "test-sub",
-                            EventName = "test-event",
-                            Subscriber = new CaptainHookContractSubscriberDto()
-                            {
-                                Webhooks = new CaptainHookContractWebhooksDto
-                                {
-                                    Endpoints = new List<CaptainHookContractEndpointDto>
-                                    {
-                                        new CaptainHookContractEndpointDto("https://webhook.site/testing", "post")
-                                    }
-                                },
-                                Callbacks = new CaptainHookContractWebhooksDto
-                                {
-                                    Endpoints = new List<CaptainHookContractEndpointDto>
-                                    {
-                                        new CaptainHookContractEndpointDto("https://webhook.site/callbacktesting", "put")
-                                    }
-                                },
-                                DlqHooks = new CaptainHookContractWebhooksDto
-                                {
-                                    Endpoints = new List<CaptainHookContractEndpointDto>
-                                    {
-                                        new CaptainHookContractEndpointDto("https://webhook.site/dlqtesting", "post")
-                                    }
-                                }
-                            }
-                        }
-                    }
-                });
-
-            result.Should().BeEquivalentTo(expected, opt => opt
-                .Excluding(info => info.SelectedMemberInfo.Name == "Authentication")
-                .Excluding(info => info.SelectedMemberInfo.MemberType == typeof(FileInfoBase)));
-
-        }
-
-
-        [Fact, IsUnit]
-        public void ProcessDirectory_JsonFileWithVars_ReturnsValidObject()
-        {
-            // Arrange
-            var jObject = GetJObjectWithVars();
-
-            // Act
-            var result = SubscribersDirectoryProcessor.GetFileVars(jObject);
-
-            // Assert
-            var expected = new Dictionary<string, Dictionary<string, string>> {
-                                {
-                                    "leviss-url", new Dictionary<string, string>
-                                    {
-                                        {"prod", "https://blah.blah.company.com/Order/Createresponse"},
-                                        {"sand", "https://blah.blah.company.test.com/v1/Order/Createresponse"}
-                                    }
-                                },
-                                {
-                                    "leviss-auth", new Dictionary<string, string>
-                                    {
-                                        {
-                                            "prod", @"{""type"":""Basic"",""username"":""abc"",""passwordKeyName"":""man--something""}"
-                                        },
-                                        {
-                                            "sand", @"{""type"":""Basic"",""username"":""def"",""passwordKeyName"":""man--other""}"
-                                        }
-                                    }
-                                },
-                                {
-                                    "sts-settings", new Dictionary<string, string>
-                                    {
-                                        {
-                                            "prod", @"{""type"":""OIDC"",""scopes"":[""any.valid.scope""]}"
-                                        },
-                                        {
-                                            "sand", @"{""type"":""OIDC"",""scopes"":[""any.valid.scope""]}"
-                                        }
-                                    }
-                                }
-                            };
-
-
-
-
-            result.Should().BeEquivalentTo(expected);
-            jObject.Should().NotContainKey("vars");
         }
 
         private static Dictionary<string, MockFileData> GetOneSampleInputFile()
@@ -329,120 +197,6 @@ namespace Platform.Eda.Cli.Tests.Commands.ConfigureEda
                 }
             };
             return mockFiles;
-        }
-
-        private static Dictionary<string, MockFileData> GetInvalidJsonMockInputFile()
-        {
-            var mockFiles = new Dictionary<string, MockFileData>
-            {
-                {
-                    "sample3.json", new MockFileData(@"<json>File</json>")
-                }
-            };
-            return mockFiles;
-        }
-
-        private static Dictionary<string, MockFileData> GetSampleInputFileWithDlq()
-        {
-            var mockFiles = new Dictionary<string, MockFileData>
-            {
-                {
-                    "sample1.json", new MockFileData(@"
-{
-  ""subscriberName"": ""test-sub"",
-  ""eventName"": ""test-event"",
-  ""subscriber"": {
-	    ""webhooks"": {
-		    ""endpoints"": [{
-			    ""uri"": ""https://webhook.site/testing"",
-			    ""authentication"": {
-				    ""type"": ""Basic"",
-				    ""username"": ""test"",
-				    ""passwordKeyName"": ""AzureSubscriptionId""
-			    },
-			    ""httpVerb"": ""post""		
-		    }]
-	    },
-	    ""callbacks"": {
-		    ""endpoints"": [{
-			    ""uri"": ""https://webhook.site/callbacktesting"",
-			    ""authentication"": {
-				    ""type"": ""Basic"",
-				    ""username"": ""test"",
-				    ""passwordKeyName"": ""AzureSubscriptionId""
-			    },
-			    ""httpVerb"": ""put""		
-		    }]
-	    },
-	    ""dlqHooks"": {
-		    ""endpoints"": [{
-			    ""uri"": ""https://webhook.site/dlqtesting"",
-			    ""authentication"": {
-				    ""type"": ""Basic"",
-				    ""username"": ""test"",
-				    ""passwordKeyName"": ""AzureSubscriptionId""
-			    },
-			    ""httpVerb"": ""post""		
-		    }]
-	    }
-    }
-}
-")
-                }
-
-            };
-            return mockFiles;
-        }
-
-        private static JObject GetJObjectWithVars()
-        {
-            return JObject.Parse(@"
-{
-  ""vars"": {
-    ""leviss-url"": { 
-      ""prod"": ""https://blah.blah.company.com/Order/Createresponse"",
-      ""sand"": ""https://blah.blah.company.test.com/v1/Order/Createresponse""
-    },
-    ""leviss-auth"": {
-      ""prod"": {
-        ""type"": ""Basic"",
-        ""username"": ""abc"",
-        ""passwordKeyName"": ""man--something""
-      },
-      ""sand"": {
-        ""type"": ""Basic"",
-        ""username"": ""def"",
-        ""passwordKeyName"": ""man--other""
-      }
-    },
-    ""sts-settings"": {     
-      ""prod,sand"": {
-        ""type"": ""OIDC"",
-        ""scopes"": [
-          ""any.valid.scope""
-        ]
-      }
-    }
-  },
-  ""subscriberName"": ""test-sub"",
-  ""eventName"": ""test-event"",
-  ""subscriber"": {
-    ""webhooks"": {
-      ""endpoints"": [
-        {
-          ""uri"": ""https://blah.blah/testing"",
-          ""authentication"": {
-            ""type"": ""Basic"",
-            ""username"": ""test"",
-            ""passwordKeyName"": ""AzureSubscriptionId""
-          },
-          ""httpVerb"": ""post""
-        }
-      ]
-    }
-  }
-}
-");
         }
     }
 }
